+++ CHANGELOG.md
@@ -6,3 +6,4 @@
   - manage model(graph)
     - manage schemas for a model
 - Add load db config from yaml
+- Add asFirstNode、asFirstEdge Methods for DataItem